Finish to refactor all exist code
